requirements.txt
Flask>=3.0,<4.0
minio>=7.2,<8.0
psycopg2-binary>=2.9,<3.0
